:new: Added conditions to commands
Simple commands can still be sent via COMM_QUEUE.put("ZMOFF"), but the
new system also allows a more sophisticated command like
COMM_QUEUE.put(["ZMOFF", ["SI?=SIMPLAY", True]]).
In thi case, "ZMOFF" will only be sent if the result of the command
"SI?" contains "SIMPLAY". The boolean in the end specifies whether the
command will only be sent if the condition is met, or if it's not met.