:x: deactivate for now
The plugin isn't really ready for production. Setting it up and keeping
it running requires too much effort for now, and in addition it's been
violating Facebook's TOS and exposing the user to the risk of losing
his/her Facebook account.

This is a nice prrof of concept, but it has no place outside my
dev-branch.